+++ example/web-app/js/ui-event.js
@@ -2,6 +2,6 @@ const eventset = require('eventset');
 
 var UIEvent = eventset.createTopic('app-ui-event');
 
-UIEvent.addEvent('toggle-list');
+UIEvent.addEvent('show-list');
 
 module.exports = UIEvent;