+++ changelog.md
@@ -2,8 +2,8 @@
 All notable changes to this project will be documented in this file
 
 ## v1.8.0
-- Non-Blocking call to event listeners event when listenr throws Error.
-- **Topic.dispatch()** accepts error callback function as third argument. 
+- Asynchronous and Non-Blocking call of event listeners, even when listenrs throws Error.
+- **Topic.addListener()** accepts error callback function as third argument. 
   By default the Error is sent to the console.
 - Rename **Eventset.Topic()** to **Eventset.createTopic()**